requirements.txt
browser-use>=0.1.40
google-generativeai>=0.3.0
playwright>=1.40.0
python-dotenv>=1.0.0
asyncio>=3.4.3
